requirements.txt
streamlit
pystac-client
boto3
python-dotenv
numpy
matplotlib
rasterio